guardian-angel-python/requirements.txt
Flask==3.0.0
Flask-PyMongo==2.3.0
pymongo==4.6.0
python-dateutil==2.8.2
gunicorn==21.2.0
packaging==23.2
requests==2.31.0
flasgger==0.9.7.1
flask_restful==0.3.9
python-dotenv==0.19.1
apscheduler==3.7.0
scikit-fuzzy==0.4.2
numpy==1.25.2
matplotlib==3.8.2
mongomock==4.0.0



